Crate resolved_ts

Resolved TS is a timestamp that represents the lower bonud of incoming Commit TS Through this timestamp we can get a consistent view in the transaction level.

To maintain a correct Resolved TS, these premises must be satisfied:

Tracing all locks in the region, use the minimal Start TS as Resolved TS.

If there is not any lock, use the latest timestamp as Resolved TS.

Resolved TS must be advanced by the region leader after it has applied on its term.

Modules

advance

cmd

endpoint

errors

metrics

observer

resolver

scanner

sinker

Structs

AdvanceTsWorker

DummySinker

Endpoint

Observer

RESOLVED_TS_ADVANCE_METHOD

Resolver

ScanTask

ScannerPool

SinkCmd

Enums

ChangeLog

ChangeRow

Error

ScanEntry

ScanMode

Task

Traits

CmdSinker

Type Definitions

BeforeStartCallback

IsCancelledCallback

OnEntriesCallback

OnErrorCallback

Result